pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "pyyaml>=6.0.3",
    "streamlit>=1.56.0",
]
